+++ earchive/earchive.csproj
@@ -115,7 +115,7 @@
     <Compile Include="gtk-gui\generated.cs" />
     <Compile Include="Loaders\ImageLoader.cs" />
     <Compile Include="MainWindow.cs" />
-    <Compile Include="Print\DocumentsPrinter.cs" />
+    <Compile Include="Print\PrintableImagesPrinter.cs" />
     <Compile Include="Print\PrintableImage.cs" />
     <Compile Include="ProductService.cs" />
     <Compile Include="Program.cs" />
